% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AoA}
\alias{AoA}
\title{Age of Aquisition Norms for English Words (AoA)}
\format{
A data frame with 43998 rows and 8 variables:
}
\source{
\url{https://osf.io/kz2px/}
}
\usage{
AoA
}
\description{
A publicly available dataset with the Test-based age-of-acquisition norms for 44 thousand English word meanings.
}
\references{
Brysbaert, M., Biemiller, A. (2017). \href{https://link.springer.com/article/10.3758/s13428-016-0811-4}{Test-based age-of-acquisition norms for 44 thousand English word meanings} In \emph{Behavior Research Methods} 49, 1520–1523.
}
\keyword{datasets}
