% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{Concrete Norms for 40 thousand English Word lemmas}
\format{
A data frame with 39954 rows and 9 variables:
}
\source{
\url{https://link.springer.com/article/10.3758/s13428-013-0403-5#MOESM1}
}
\usage{
concrete
}
\description{
A publicly available dataset with the Test-based age-of-acquisition norms for 44 thousand English word meanings.
}
\references{
Brysbaert, M., Warriner, A.B. & Kuperman, V. (2014). \href{https://link.springer.com/article/10.3758/s13428-013-0403-5}{Concreteness ratings for 40 thousand generally known English word lemmas} In \emph{Behavior Research Methods} 46, 904–911.
}
\keyword{datasets}
