% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_info}
\alias{dict_info}
\title{Dictionary Info}
\format{
An object of class \code{data.frame} with 12 rows and 5 columns.
}
\usage{
dict_info
}
\description{
A summary table with the names of dictionaries included in the package, the type of tokens which the dictionary accepts (tokens/lemmas), the type of the dictionary (norms or "simple" like LIWC), the columns of the dataset to use in the text analysis and the default names given to the output of each measure in the dictionary.
}
\keyword{datasets}
