% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evlex}
\alias{evlex}
\title{The Evaluative Lexicon}
\format{
A data frame with 1541 rows and 4 variables.
}
\usage{
evlex
}
\description{
A publicly available dataset with the valence, extremity and emotionality of word lemmas. From the TATE package.
}
\references{
Rocklage, M. D., & Fazio, R. H. (2015). The Evaluative Lexicon: Adjective use as a means of assessing and distinguishing attitude valence, extremity, and emotionality. In \emph{Journal of Experimental Social Psychology}, 56, 214-227.
}
\keyword{datasets}
