% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humor}
\alias{humor}
\title{Humor Norms}
\format{
A data frame with 4997 rows and 16 variables.
}
\usage{
humor
}
\description{
A publicly available dataset with humor scores of word lemmas. From the TATE package.
}
\references{
Engelthaler, T., & Hills, T. T. (2018). Humor norms for 4,997 English words. In \emph{Behavior research methods}, 50(3), 1116-1124.
}
\keyword{datasets}
