% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lancaster}
\alias{Lancaster}
\title{The Lancaster Sensorimotor Norms}
\format{
A data frame with 39707 rows and 45 variables:
}
\source{
\url{https://osf.io/7emr6/}
}
\usage{
Lancaster
}
\description{
A publicly available dataset with the multidimensional measures of perceptual and action strength for 40,000 English words.
}
\references{
Lynott, D., Connell, L., Brysbaert, M. et al. (2020). \href{https://link.springer.com/article/10.3758/s13428-019-01316-z}{The Lancaster Sensorimotor Norms: multidimensional measures of perceptual and action strength for 40,000 English words.} In \emph{Behavior Research Methods} 52, 1271–1291.
}
\keyword{datasets}
