% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{liwcalike_dicts}
\alias{liwcalike_dicts}
\title{A function for text analysis with simple count dictionaries}
\usage{
liwcalike_dicts(now_toks, now_dict)
}
\arguments{
\item{now_toks}{The tokens from the text pre-processing.}

\item{now_dict}{The simple ("LIWCalike") dictionary to use for analysis. Can be fitting dictionaries from the package or any dictionary with the format of a quanteda::dictionary() (e.g. a dictionary of words that represent negative emotion).}
}
\description{
This function performs an analysis on pre-processed text, calculating the proportions of words per dictionary in the text.
This function isn't limited to dictionaries in the package and can be used with additional custom dictionaries (as opposed to the analyze_text function).
}
\examples{

proc_toks <- preproc_text(data.frame(text="I wasted time, and now doth time waste me."))
liwcalike_dicts(proc_toks,AFINN)


}
\keyword{dictionary}
\keyword{lexicon,}
