% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MPD}
\alias{MPD}
\title{The Mind Perception Dictionary (MPD)}
\format{
An object of class dictionary2 of length 5.
}
\source{
\url{https://www.adamwaytz.com/measures-and-materials/the-mind-perception-dictionary}
}
\usage{
MPD
}
\description{
A quanteda dictionary object with A publicly available dictionary that can assess frequency of words related to mind perception.
}
\references{
Schweitzer, S., & Waytz, A. (2021). \href{https://psycnet.apa.org/fulltext/2020-96445-001.pdf}{Language as a window into mind perception: How mental state language differentiates body and mind, human and nonhuman, and the self from others.} In \emph{Journal of Experimental Psychology: General} 150(8), 1642.
}
\keyword{datasets}
