% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MPVN}
\alias{MPVN}
\title{The Mental-Physical Verb Norms (MPVN)}
\format{
A data frame with 250 rows and 6 variables:
}
\source{
\url{https://www.adamwaytz.com/measures-and-materials/the-mind-perception-dictionary}
}
\usage{
MPVN
}
\description{
A publicly available dataset with the Mental-Physical Verb Norms (MPVN) scores of word lemmas to measure mental state attribution.
}
\references{
Orr, R. I., & Gilead, M. (2022). \href{https://link.springer.com/article/10.3758/s13428-022-01911-7}{Development and validation of the Mental-Physical Verb Norms (MPVN): A text analysis measure of mental state attribution} In \emph{Behavior Research Methods} 1-21.
}
\keyword{datasets}
