% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{norm_dicts}
\alias{norm_dicts}
\title{A function for text analysis with norm based dictionaries}
\usage{
norm_dicts(now_toks, now_dict, column, word_column = "Word")
}
\arguments{
\item{now_toks}{The tokens / lemmas from the text pre-processing.}

\item{now_dict}{The norms dictionary to use for analysis. Can be fitting dictionaries from the package or any norm dictionary with the format of a data.frame with one column 'Word' which includes tokens / lemmas and other columns providing norm based scores (e.g. concreteness level per word).}

\item{column}{The column/columns in the dictionary to be used in the analysis.}

\item{word_column}{The column in the dictionary which is the index of tokens / lemmas Defaults to 'Word'.}
}
\description{
This function performs an analysis on pre-processed text, calculating the levels of the selected norm-based dictionary in the text.
This function isn't limited to dictionaries in the package and can be used with additional custom norm dictionaries (as opposed to the analyze_text function).
}
\examples{
proc_toks <- preproc_text(data.frame(text="Uneasy lies the head that wears the crown."))
norm_col <-  dict_info$norm_column[[which(dict_info$dict_names=="concrete")]]
result <- norm_dicts(proc_toks,now_dict=concrete,column=norm_col)
colnames(result)[2] <- dict_info$out_norm_column[[which(dict_info$dict_names=="concrete")]]
result

proc_lems <- preproc_text(data.frame(text="Uneasy lies the head that wears the crown."),dict_entry = 'lemmas')
norm_col <-  dict_info$norm_column[[which(dict_info$dict_names=="VAD")]]
result <- norm_dicts(proc_toks,now_dict=VAD,column=norm_col)
colnames(result)[2:4] <- dict_info$out_norm_column[[which(dict_info$dict_names=="VAD")]]
result
}
\keyword{dictionary}
\keyword{lexicon,}
