% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRC}
\alias{NRC}
\title{NRC Word-Emotion Association Lexicon}
\format{
An object of class dictionary2 of length 10:
}
\source{
\url{http://saifmohammad.com/WebPages/AccessResource.htm}
}
\usage{
NRC
}
\description{
A quanteda dictionary object with the English version of the NRC Word-Emotion Association Lexicon. The association of words with eight emotions (anger, fear, anticipation, trust, surprise, sadness, joy, and disgust) and two sentiments (negative and positive) manually annotated on Amazon's Mechanical Turk. From quanteda.dictionaries.
}
\references{
Mohammad, S. & Turney, P. (2013). \href{https://arxiv.org/abs/1308.6297}{Crowdsourcing a Word-Emotion Association Lexicon.} In \emph{Computational Intelligence}, 29(3), 436–465.

Mohammad, S. & Turney, P. (2010). \href{https://dl.acm.org/citation.cfm?id=1860635}{Emotions Evoked by Common Words and Phrases: Using Mechanical Turk to Create an Emotion Lexicon.} In \emph{Proceedings of the NAACL-HLT 2010 Workshop on Computational Approaches to Analysis and Generation of Emotion in Text}, June 2010, LA, California.
}
\keyword{datasets}
