% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{preproc_text}
\alias{preproc_text}
\title{A function for pre-processing text for analysis}
\usage{
preproc_text(data, text_column = "text", dict_entry = "tokens")
}
\arguments{
\item{data}{The data set with texts to analyze.}

\item{text_column}{The name of the column in data with the texts. Defaults to 'text'.}

\item{dict_entry}{Whether to split text into tokens or lemmas. Defaults to 'tokens'.}
}
\description{
This function pre-processes a text for analysis by tokenizing or lemmatizing the text.
}
\examples{
preproc_text(data.frame(text="For sweetest things turn sourest by their deeds"))

preproc_text(data.frame(text="For sweetest things turn sourest by their deeds"),dict_entry='lemmas')
}
\keyword{dictionary}
\keyword{lexicon,}
