% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Prevalences}
\alias{Prevalences}
\title{Word prevalence norms for 62,000 English lemmas}
\format{
A data frame with 61855 rows and 5 variables:
}
\source{
\url{https://osf.io/g4xrt/}
}
\usage{
Prevalences
}
\description{
A publicly available dataset with the prevalence measures of 62 thousand english lemmas. Word prevalence refers to the number of people who know the word. The measure was obtained on the basis of an online crowdsourcing study involving over 220,000 people.
}
\references{
Brysbaert, M., Mandera, P., McCormick, S. F., & Keuleers, E. (2019). \href{https://link.springer.com/article/10.3758/s13428-018-1077-9}{Word prevalence norms for 62,000 English lemmas.} In \emph{Behavior Research Methods} 51(2), 467-479.
}
\keyword{datasets}
