% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{threat}
\alias{threat}
\title{The threat Dictionary}
\format{
An object of class dictionary2 of length 1.
}
\source{
\url{https://www.michelegelfand.com/threat-dictionary}
}
\usage{
threat
}
\description{
A quanteda dictionary object with A publicly available dictionary that indexes threat levels from texts.
}
\references{
Choi, V. K., Shrestha, S., Pan, X., & Gelfand, M. J. (2022). \href{https://www.pnas.org/doi/abs/10.1073/pnas.2113891119}{When danger strikes: A linguistic tool for tracking America’s collective response to threats} In \emph{Proceedings of the National Academy of Sciences} 119(4), 19.
}
\keyword{datasets}
