% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VAD}
\alias{VAD}
\title{The Evaluative Lexicon}
\format{
A data frame with 13915 rows and 65 variables:
}
\usage{
VAD
}
\description{
A publicly available dataset with the values of valence, arousal and dominance of word lemmas. From the TATE package.
}
\references{
Warriner, A. B., Kuperman, V., & Brysbaert, M. (2013). Norms of valence, arousal, and dominance for 13,915 English lemmas. In \emph{Behavior research methods}, 45(4), 1191-1207.
}
\keyword{datasets}
